multiple consumers
on adding multiple consumers , we had to add a mutex & a waitGroup
The mutex prevents 2 consumers from trying to modify the sentence at the same time because this involves reading and writing a new value to sentence which can lead to a race condition.

The waitGroup is necessary because we can no longer consider the execution to be complete when only one consumer has finished , we must wait till all the consumers that we have instantiated have finished.